requirements-grounding.txt
transformers>=4.45,<5
torch>=2.3
Pillow>=10
